Add tags for nix run shell commands (#41)
Prefer stable nixpkgs for normal usage, so that we hit the cache
more often (this should be paired with an increase in nix store ttl
now that I have a better laptop). Also use unstable nixpkgs as
a secondary option so that we have a better chance of hitting the
nix foundation binary cache.