Add Kanata keybinding daemon to darwin systems (#86)
Kanata allows for keyboard and keybind manipulation at the input
device level. This makes it suitable for controlling input to all
other applications and daemons, for example the shkd hotkey daemon,
which requires that any manipulations (such as adding a super key)
occur below it in the keyboard stack.

This allows better system-wide keybinds, sets me up for a return to
the shkd hotkey daemon as well as yabai, and opens the door for more
complex keyboard manipulations in the future.