remove tuffy builds, replace with generic nix flake check
Currently, tuffy will not build on GitHub Actions, as it is
an aarch64 machine and they do not have a public aarch64 platform
yet.

Instead, just run a nix flake check and hope that it is sufficient.
I suspect not, but it will be difficult to find an acceptable
alternative.