fix: rekey tuffy secrets for tuffy (#163)
`sops` automatically assumes that the keys already present in a file
should continue to be used if secrets get updated. This is unintuitive
if you update .sops.yaml but do not run `sops updatekeys <file>` on
a copied (or rekeyed) file.

On a more secure system, the tuffy keys should be rotated. We choose
not to for simplicity.