alacritty: allow new windows
Previously, we disallowed new windows along with new tabs,
because for some reason a high window switching time was
causing me to cmd+T in alacritty, which at the time may have
caused a crash.

This is not an issue for new windows, so I shouldn't block them.
I'd like to use them to create portable windows instead.