linux-builder: set core limit
I'd like to keep running a linux builder on Mac OS, so that I
have offline linux build support. (Though I am sufficiently
internet-connected that it may be time to just bite the bullet
and set up a real builder, and deprecate this.)

Limit it to two cores so it cannot have too much of an effect
on system resources.