tailscale: automatic secrets rotation
I always forget to do secret rotation. This is most
critical with Tailscale auth keys, because machines
without fresh keys will drop off the network often.

Additionally, I think it is good discipline to have
means of automatically refreshing secrets.

There is a drawback to this system: old secrets are
not revoked. In the naive implementation, we'd need
to retrieve the old secrets to revoke them, and I'd
prefer that there is no one key that can access the
written secrets. An alternative architecture may be
to query authkeys and revoke anything older than 30
days or so, only using the Tailscale API.

Changelist:

* Pull tailscale secrets into their own files
  so other secrets don't need to be rewritten.

* Make tailscale secrets write-only.

* Add a script to rotate these secrets using
  Tailscale's get-authkey and sops.