tmux: don't intercept escapes for so long
https://man.openbsd.org/cgi-bin/man.cgi/OpenBSD-current/man1/tmux.1?query=tmux&sec=1#escape-time

https://unix.stackexchange.com/questions/608142/whats-the-effect-of-escape-time-in-tmux

Because tmux is waiting to see if a sequence follows the escape key,
it waits before sending it to the running program. By default this
is 500ms, which is _very_ noticable. Especially in an editor, I am
often planning my next movement before I finish an insertion, and
so I start typing before the 500ms ends and tmux cancels the escape.
This is annoying!

I can't remove it outright, but I can guarantee that most of my
computer use is on my local machine and even on remote machines I
can guarantee pretty solid latency. A direct connection to marlon
in Helsinki is about 130ms, which could be problematic. I don't know
much about the SSH protocol but I assume there is not a round-trip
between keys, which would imply that the time between keys is
generally less than that time.

Therefore, I turn it down to 20ms. If I have trouble with modifiers
or odd keys, I should turn it back up.