kernel: prefer swap on zram
Swap into compressed pages via zram is now viable.
It turns out that memory compresses pretty well, and that
spending that CPU time compressing and decompressing is
comparable to a write-out to disk in terms of latency.

By default, allow up to 100% memory in ZRAM. Assuming a 2x
compression ratio, this results in 50% of memory being taken
up by swap for a total of 150%.

This doesn't remove disk swap. I'll consider it!
Especially on the high-RAM machines.