platforms/hetzner: re-enable ipv4
Setting a static configuration for ipv6 seems to disable DHCP for
ipv4 (which is enabled by default otherwise). This means that the
system is no longer able to access ipv4-only sites for the time
being.

Fix this by explicitly enabling DHCP on the physical interface,
which will automatically configure ipv4.