misc: speed optimizations
Using systemd initrd allows for measuring separate kernel/init
boot times. It also seems to add a modest speedup.

== DHCP ==

I set noarp on dhcpcd because we don't expect to every be
in a situation where we receive an address assigned to
another machine.

I also set rapid_commit on dhcpcd because we don't expect
multiple DHCP servers.

Probably there is room for improvement by setting static
IPs where they are known and not relying on DHCP at all.

== Kernel ==

This is a virtualized server, it doesn't do WIFI. I'm not
sure how or why WIFI modules are loading. Blackless
cfg80211 and rfkill, because we don't need them.