keycloak: init
Add keycloak for authentication.

This allows auth.tuckershea.com to be publicly accessible via marlon,
and the admin panel to be privately accessible via spencer.

The tags might need some reworking. Maybe it is worth putting
keycloak on its own port for easier firewalls/ACLs. I think it will
take me some time to learn about how this is architected
in practice, but for the moment this is sufficient.