minecraft-server abtech: init
Adds a fabric server with quite a few mods!

I don't intend to update without substantial work so I don't
mind hardcoding all the mods here. One day I may regret it!

Set up for a pretty specific seed:
https://map.jacobsjo.eu/?seed=2430&datapacks=modrinth:terralith,modrinth:tectonic,modrinth:dungeons-and-taverns

and additionally for a small worldborder. When the server starts,
the worldborder is set, chunks can be generated up to it by chunky,
and then a map can be rendered with dynmap. The map is then exposed
via nginx.

Even though it is the only server on this particular machine,
I do have it running behind the gate proxy. Mostly because I wanted
to get it to work while I had the time. It's configured here as
just a direct proxy; it does not intercept any commands.

NOTE: dynmap is configured at runtime! The only major changes I made
were threading and setting it to use sqlite backing store.