machines/makoto: init
Add makoto, a NAS which I've taken over with Nix. It will serve
solely as an offsite backup server, and otherwise will have as
minimal of a configuration as possible.

The hardware is a UGREEN NASync DXP2800. It is very easy to
install another OS on: disable "Watchdog" in BIOS and just use
a usb stick.

There are two 10tb hard disks in a ZFS mirror.